refactor: isolate fetchPokemonDescriptions() in searchPokemons.ts
