+++ react-components/src/API/index.ts
@@ -1,5 +1,4 @@
-import { fetchPokemonDescriptions } from './fetchPokemonDescriptions';
 import { fetchPokemonNames } from './fetchPokemonNames';
 import { searchPokemons } from './searchPokemons';
 
-export { fetchPokemonNames, fetchPokemonDescriptions, searchPokemons };
+export { fetchPokemonNames, searchPokemons };
